prevent disk caching, move helmet config to helper file
